requirements.txt
Django==1.11.10
django-webpack-loader==0.6.0
Fabric3==1.14.post1